requirements.txt
discord.py
aiohttp
requests
prompt-toolkit
Flask
gunicorn
python-dotenv
